mcg controller

+++ database/migrations/2020_11_03_185954_create_experiencia_laborals_table.php
@@ -41,7 +41,7 @@ class CreateExperienciaLaboralsTable extends Migration
             $table->string("cargo_comision")->nullable();
             $table->string("funcion_pricipal")->nullable();
             $table->string("especifique_otro_sector")->nullable();
-            $table->double("salario_mensual_neto")->nullable();
+            $table->integer("salario_mensual_neto")->nullable();
             $table->unsignedBigInteger("respuesta_proveedor_id")->nullable();
             $table->foreign("respuesta_proveedor_id")->references("id")->on("respuestas");            $table->timestamps();
             $table->softDeletes();
